* commit '6651209caef29492858539b56aa8202482598f27':
  make_fs(): don't hardcode argv index
  ploop tool: usage formatting fixes
  ploop tsnapshot(): fix usage
  Fix POWER compilation wrt __u64 printf/scanf
  ploop_set_max_delta_size(): fix definition
  get_max_ploop_size(): change type of last arg
  mount.ploop: check if mtab is not symlink

+++ lib/gpt.c
@@ -225,8 +225,8 @@ static int update_gpt_partition(int fd, const char *devname, __u64 new_size512,
 
 	if (new_size512 > size) {
 		ploop_err(0, "Unable to resize GPT partition:"
-				" incorrect parameter new_size=%llu size=%lu",
-				new_size512, (long)size);
+				" incorrect parameter new_size=%" PRIu64 " size=%lu",
+				(uint64_t)new_size512, (long)size);
 		return SYSEXIT_PARAM;
 	}
