+++ lib/ploop.c
@@ -246,7 +246,7 @@ static int default_fmt_version(void)
 	return ploop_is_large_disk_supported() ? PLOOP_FMT_V2 : PLOOP_FMT_V1;
 }
 
-static int get_max_ploop_size(int version, unsigned int blocksize, __u64 *max)
+static int get_max_ploop_size(int version, unsigned int blocksize, unsigned long long *max)
 {
 	switch(version) {
 	case PLOOP_FMT_V1:
@@ -288,7 +288,7 @@ int ploop_get_max_size(unsigned int blocksize, unsigned long long *max)
 static int do_check_size(unsigned long long sectors, __u32 blocksize, int version,
 		__u64 max_fs_size)
 {
-	__u64 max;
+	unsigned long long max;
 
 	if (version == PLOOP_FMT_UNDEFINED)
 		return 0;
@@ -1679,7 +1679,7 @@ static int create_ploop_dev(int minor)
 	return 0;
 }
 
-static int set_max_delta_size(int fd, __u64 size)
+static int set_max_delta_size(int fd, unsigned long long size)
 {
 	/* Set max delta size for the last added (top) delta */
 	ploop_log(0, "Setting maximum delta size to %llu sec", size);
@@ -2709,9 +2709,9 @@ int ploop_resize_image(struct ploop_disk_images_data *di, struct ploop_resize_pa
 				available_balloon_size = balloon_size + (fs.f_bfree * B2S(fs.f_bsize));
 				if (available_balloon_size < new_balloon_size) {
 					ploop_err(0, "Unable to change image size to %lu "
-							"sectors, minimal size is %llu",
+							"sectors, minimal size is %" PRIu64,
 							(long)new_fs_size,
-							(blocks - available_balloon_size - reserved_blocks));
+							(uint64_t)(blocks - available_balloon_size - reserved_blocks));
 					ret = SYSEXIT_PARAM;
 					goto err;
 				}
@@ -3392,7 +3392,7 @@ int ploop_get_spec(struct ploop_disk_images_data *di, struct ploop_spec *spec)
 	return ret;
 }
 
-int ploop_set_max_delta_size(struct ploop_disk_images_data *di, __u64 size)
+int ploop_set_max_delta_size(struct ploop_disk_images_data *di, unsigned long long size)
 {
 	char conf[PATH_MAX];
 	char dev[64];
